options.py
VERBOSE_MODE = True

from dotenv import find_dotenv, load_dotenv

from langchain import HuggingFacePipeline, PromptTemplate, LLMChain
from transformers import AutoTokenizer, pipeline
from langchain.llms import GPT4All
from langchain.callbacks.streaming_stdout import StreamingStdOutCallbackHandler

import torch
import requests
import os

import streamlit as st



