requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.8
decorator==5.1.1
einops==0.6.1
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpt4all==0.3.4
greenlet==2.0.2
grpcio==1.54.2
h5py==3.8.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
jax==0.4.12
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.12.0
langchain==0.0.202
langchainplus-sdk==0.0.10
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.3
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scipy==1.10.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.16
streamlit==1.23.1
sympy==1.12
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.30.2
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
validators==0.20.0
Werkzeug==2.3.6
wrapt==1.14.1
yarl==1.9.2
zipp==3.15.0
